third_party/WebKit/LayoutTests/fast/mediarecorder/BlobEvent-basic.html

<!DOCTYPE html>
<script src=../../resources/testharness.js></script>
<script src=../../resources/testharnessreport.js></script>
<script>

test(function() {
  var array = new Uint8Array([0x70, 0x71, 0x72, 0x73]);
  var blob = new Blob([array]);
  var blobEvent = new BlobEvent('BlobEvent', {data : blob});

  var reader = new FileReader();
  reader.addEventListener("loadend", function() {
    // |reader.result| contains the contents of blob as an ArrayBuffer.
    var outputArray = new Uint8Array(reader.result);
    assert_array_equals(array, outputArray)
  });
  reader.readAsArrayBuffer(blob);

}, 'check BlobEvent creation and content management');

</script>
